Support different scalar types explicitly
Operations such as * and / now (as well as + and - for consistency)
support arbitrary types. Thus the value is passed as is to
the actual scalar operator.

This fixes the problem when multiplication by double < 1
resulted in 0 because the value was converted to 0
in the function call to fit the operator (function)
parameter. The bug was introduced in 87eda93dc
by changing double to Number in these functions.
Now no conversion is done when not needed and
a proper conversion is done as part of the computation
when necessary.

Bahavior for multiplication of raster with scalar and two raster added.