Avoid creating a raster in simulation
Now caller needs to pass an existing dispersers raster to the simulation.
The con that it needs to exist and user needs to keep track of it,
but the pro is that it can be examined and used further or even modified
in between steps of simulation.

Consequently the dispersers are no longer a hidden detail, but part of what is
exposed to the caller.

The motivation is the avoiding of raster creation in simulation which makes
the simulation less dependent on particular raster class behavior,
in this case, a specific constructor (int, int) is no longer needed.