Improve Simulation class API and documentation (#44)
Simulation constructor does not take a raster anymore just to grab the size from it and instead it
requires rows and columns as parameters.
This comes also with documentation for Simulation function parameters.

Simulation has now an optional template parameter for index type. This seemed natural for the
constructor parameters and member variables since it is bind to the raster classes,
but the relationship with type (int) used internally and in dispersal kernel may need to be revised
since they are using specific type and any generic type needs to be signed for dispersal kernel
mechanism to work properly, so, for example,
the typical C++ standard library (STL) index (size_t) is not usable.

A lot of lines are touched by renaming the member/private variables (from width-height to rows-cols terminology and order) and adding the trailing underscore.