Add exposed into possile steps
Exposed are tracked similarly to moratlity in a vector of rasters
where the position determines the age. The size of the vector indicates
time from the start of the simulation and the vector needs to be increased
outside of the Simulation each step until it reaches latency period (in steps)
plus one.

The result of SI is identical with SEI with latency period of zero steps.

A new wrapper function is provided to wrap the current disperse call combined
with the E to I step of SEI model.