A simple deteministic kernel for tests of simulation (#53)
This adds a new simple non-random kernel and test which is using this kernel to test Simulation without the randomness of the kernel thus having a predictable outcome which can be tested for exact values. The output values are computed by hand, but to match them another stochastic part needs to be disabled and that's the dilution effect which is minimalized by having a lot of hosts, so the ration of S to all does not change that much.