Expose the deterministic neighbor kernel in the rest of the API
The DeterministicNeighborDispersalKernel was available, but only when the objects are directly constructed by the programmer.
Now the kernel is part of the kernel API alongside other kernels.

The SwitchDispersalKernel constructor now has an additional parameter, but it it optional,
so the API is the same as before if the DeterministicNeighborDispersalKernel is not used
(however, in that case, it needs to be not allowed in the user interface because
the string to kernel enum function will not complain).