Add a higher level interface to the simulation
This adds a higher level interface to the simulation which connects several lower level
objects into one API to unify and simplify use accross the interfaces.

The name for this API is Model (pops::Model) as in PoPS Model while the class Simulation
is the lower level engine making the main part of model mechanics happen.

This depends on several PRs which are not yet merged at this point.