Add scheduling to config
Now config includes some of the scheduling logic because that's something
which needs to be done and checked as a model run is being prepared
to get the sizes of various vectors right and check the inputs,
i.e., doing this in the Model class is too late.

Since some of the values need to be computed, a getters with a check of a state
are used. The implementation needs to check the state, but the usage is more flexible.