Add a higher level interface to the simulation (#70)
This adds a higher level interface to the simulation which connects several lower level
objects into one API to unify and simplify use across the interfaces.

The name for this API is Model (pops::Model) as in PoPS Model while the class Simulation
is the lower level engine making the main part of model mechanics happen.

The values are initialized to the currently used defaults, so no to features/complexity, yes to stochasticity.

Config includes some of the scheduling logic because that's something
which needs to be done and checked as a model run is being prepared
to get the sizes of various vectors right and check the inputs,
i.e., doing this in the Model class is too late.

Since some of the values need to be computed, a getters with a check of a state
are used. The implementation needs to check the state, but the usage is more flexible.