Use CMake for build and a common project structure (#59)
This includes:

* Introduction of CMake build and test system.
* Move source code files to commonly used directories.
* Adds CMake build and test to GitHub Actions CI.
* Update to latest clang-format-lint-action from a branch in a fork.

This is needed to integrate nicely into PyPoPS which is using CMake and can make use of CMake for dependencies.

Headers are in an include directory with given directory name for the library. Headers in tests are included as system headers. Tests are separated in a dedicated directory.

The current make system is still in place and updated. The files themselves are not modified. The paths to header files are changed and pops is now more promoted as the directory to include from as <pops/...hpp> but unless CMake is used, it is up to the user, so the current rpops and r.pops.spread builds can still choose.

Makefile is updated to work with the new structure but marked as obsolete. Readme updated for CMake and usage from other projects.