Rename this project from PoPS to PoPS Core
We use PoPS as a name for the whole project and it makes sense to have also rpops and r.pops.spread
under the name PoPS. This library is just one part of it.

However, it is the core of PoPS, so calling it PoPS Core in text and pops-core elsewhere.
Keeping pops (and PoPS) in the code because it is the PoPS interface/library in C++.