Add movement of pest based on number (#127)
Pest can now leave a cell if there is too much of pests at one cell. The movement is driven by the anthropocentric (long) dispersal kernel.

Uses overpopulation movements rather than pest movements and pest rather than units (of pests) as terminology. Using percentage in the API because that's what's happening and in the naming since it is more specific than ratio.

Slightly different tests for Model and for Simulation are both just the minimal setup to test the overpopulation.

Closes #125.