Use scaled natural kernel for overpopulation (#135)
Instead of (mis)using antropogenic kernel as a long distance kernel for pest overpopulation movement, scale (multiply) the scale (distance) parameter of natural kernel using a coefficient.

This adds one new config variable to modify the scale parameter of radial and deterministic kernels. It is meant as modification, so it is called coefficient [of scale] (`leaving_scale_coefficient`) and it defaults to 1. "Coefficient of scale" is not ideal, but "scale of scale" would not be great either.

Additionally, this moves most of the kernel creation to separate documented functions. Although the duplication between the functions would be unnecessary difficult to address, the similarities between the three kernels (natural, anthro, overflow) are now more clear than in the original intermingled code and easier to keep in sync if needed. The kernels are now created only when needed (i.e., with limited scope and potentially not for every step) since they are really not only local to the run step function, but needed only for one function call.

The new functions are protected and the existing attributes are now protected instead of private because for practical reasons, similarly to other places, we consider inheritance a case when user of a class wants to deal with internals of the object.

Finally, when config is not properly used, the new behavior is that it now consistently fails while before it was sometimes passing. In the config, we now set more defaults to ensure consistent results (that is throwing an exception when not set because the default and wrong 0 is used).

Additionally, invalid radial kernel message code is simplified and the deterministic kernel now documents the use of the dispersal reference and how invalid arguments are handled.