Use virtual functions instead of if-else in switch kernel
Experimental code of a more dynamic system for handling runtime switching between multiple kernels.
Instead of if-else statements in operator(), different classes are created in the kernel creation functions.
Switching in the operator() happens using virtual table rather than if-else in the switch kernel's operator()
and the actual kernels is access though a pointer (to an abstract base class).

This may make handling multiple kernels more flexible or easier and a variation of this seems to be needed for supporting multiple network kernels.