Remove node file from network
This removes the need to load the network as a file with nodes and another file with segments. Now, only the file with segments is needed since it anyway includes nodes and their coordinates.

There is a slight speed up, but only very small because the node file was small comparing to segments file and the more complex data structures needs to be filled anyway.

However, this simplifies the inputs and required data.