Support every_step (time_step) as frequency in scheduling (#147)
* Support time_step as frequency in scheduling
* Support every_step and time_step as synonyms
* Add tests for every_step and time_step
* Remove todo tracked in a separate issue

Any simulation unit + n combination will now work with output_frequency == "time_step". The logic is that we know what is the step in the scheduler, so scheduling every step is just schedule every nth step with n = 1.

This was originally implemented in a limited form in rpops code. Now rpops, can simply rely on the library.