Replace time and speed by distance (#161)
Getting the average (or any) speed for trains is possible, but difficult. However, the speed requires also minimum and maximum trip time which seems to be quite problematic to get or estimate. The same model logic is now better represented by distance which is also easier to explain in terms of minimum and maximum trip distance (versus min and max in combination with average speed). Therefore:

* Replace time by distance in computations.
* Remove speed completely.

Note that the network test with exact destination cells stays the same because cost per cell is still 1 in that test.

One of the original ideas behind the speed was to model how the pest can spread more easily when a train slows down or is standing in a station. With better train data, speed and time might be a useful concept again. Alternatively, with lack of data, but strong need for higher spread in nodes, the spatial component of segments could be removed completely and the dispersers would jump only between nodes.