Snap points to the closest node in network
Optionally, instead of ending in somewhere in the segment, the network travel ends at the start or end node
of a segment (whichever is the closest).

To support the snapping based on distance, each segment now has explicit total cost (distance), so Segment and SegmentView
need to be classes, not just using.

Having the total cost (distance) for segment available allows the travel function to not iterate over a segment when its cost
is lower than the remaining distance (cost) to travel (spend).

This contains only the changes in the network itself.