Use citation info specific to PoPS Core
The previous file was copied from r.pops.spread with the idea that the files will be shared, i.e., will be the same. However, with specific set of DOIs for each repo, it makes more sense now to have a specific citation file for each repo. The file still points people to the readme as the primary source of information where we give suggestions and paper citations (this could be, at least partially in the citation file too, but in the readme, it is perhaps friendlier). The main purpose for the file is that Zendo uses that to build the DOI record and the citation. Additionally, GitHub links this file from the About section.