Add probability-based stepping to a node in network (#168)
This adds step (i.e., jump, teleport) from one node to another based on probability assigned to an edge.

The code it uses is completely different, so a new function step is now available alongside of the existing travel function.

Node to go to is selected based on edge probability in the new step function. Ignoring of nodes, implemented for the travel function, is not implemented for the step function (or not yet) because that would change the relative probabilities.

Edges have probabilities, but they don't have costs. Probability can be any range (it is relative) but it cannot be negative. Edge probabilities are determined from the input CSV.

Network type is now driven by a single string variable with three values. Using current names network_type = snap, step, travel.