Remove spatial index from computation of quarantine areas bboxes (#189)
The quarantine constructor cannot compute bboxes of quarantine areas by looping over suitable cells because the bboxes would be different then bboxes of the actual areas. This also avoids a segfault if width and height of the areas are 0.

No updates to quarantine test are needed since suitable cells cover entire area, so the change does not impact the test result.

Fixes #146. Reverts small part of #120 (add spatial index).