Use actions in model directly (#203)
This uses actions, host pool and pest pool in the model directly without using the Simulation class.

To remove Simulation completely, it also removes soils and initialization of Simulation from Model. Soils is passed to spread action now.

This also adds more details for the report in the reduced stochasticity test.