Add to suitable cells only when not already there (#213)
The original code adds new row, col pair to the suitable cells index when row, col matches row, col in the suitable cells. However, it should add the new cell to the list when it was not found in the list. (The list of suitable cells acts as a set.)

Additionally, search in the vector was originally done using a loop and now it is done using a standard library function find wrapped by our container_contains function.