Add multiple networks
The new MultiNetwork class works with one or more networks. The code structure is changed in a way that network can be switched statically (as a template parameter) instead of using the orinal network class but having index as its template parameter. This means that model does not know about the number of networks (or what network is actually used). The kernel newly does not know any details about the network and the network itself handles the different movements (which requires the distribution to be instantiated for every call because of constness).

Test for one network is included and configuration is available, but both need more work.