Improve the basic network test
Update the test to reflect the changes in dff7717a2f99aae554004f40f902b743395531c5 which replaced time and speed by distance. The distance is in map units not in cells, so it needs to change for this test.

The test also sets up the test data better and tests all specific values in addition to tests of general requirements.