Explicitly disable mortality in host pool (#231)
Instead of relying on size of the mortality vector, ignore the provided data completely when mortality is explicitly disabled. This provides greater flexibility on what rpops and Rcpp can do (#230).

Besides adding yet another config value to the HostPool, add also a constructor which takes a Config object, so that all values can be taken from there. Switch to that in Model and tests when HostPool is used directly, but keep using the old, but updated constructor in tests which don't use config and in Simulation because the Simulation and its tests are not using Config.

Deterministic tests had wrong mortality config which did not surfaced until now when it is used inside the host pool.